requirements.txt
behave==1.2.6
parse==1.19.0
parse-type==0.6.0
PyHamcrest==2.0.4
six==1.16.0
